requirements.txt
selenium
pillow
pandas
tqdm
user_agent
webdriver-manager